remove Component suffix from AutoComplete and move assets to a folder
