+++ docs/content/components/autocomplete.md
@@ -21,7 +21,7 @@ Use AutoComplete to populate input values from server search results.
 <Example src="<auto-complete src='/users/search' for='user-popup' class='position-relative'>  <input name='input' type='text' class='form-control '></input>    <ul id='user-popup' class='autocomplete-results '>    <li role='option' data-autocomplete-value='value' aria-selected='true' class='autocomplete-item '>      Selected</li>    <li role='option' data-autocomplete-value='value' class='autocomplete-item '>      Not selected</li></ul></auto-complete>" />
 
 ```erb
-<%= render(Primer::AutoCompleteComponent.new(src: "/users/search", id: "user-popup", position: :relative)) do |c| %>
+<%= render(Primer::AutoComplete.new(src: "/users/search", id: "user-popup", position: :relative)) do |c| %>
   <% c.input(type: :text, name: "input") %>
   <% c.results do %>
     <%= render(Primer::AutoCompleteItemComponent.new(selected: true, value: "value")) do |c| %>
@@ -39,7 +39,7 @@ Use AutoComplete to populate input values from server search results.
 <Example src="<auto-complete src='/users/search' for='user-popup' class='position-relative'>  <input name='input' type='text' class='form-control '></input>    <ul id='user-popup' class='autocomplete-results my-custom-class '>    <li role='option' data-autocomplete-value='value' aria-selected='true' class='autocomplete-item '>      Selected</li>    <li role='option' data-autocomplete-value='value' class='autocomplete-item '>      Not selected</li></ul></auto-complete>" />
 
 ```erb
-<%= render(Primer::AutoCompleteComponent.new(src: "/users/search", id: "user-popup", position: :relative)) do |c| %>
+<%= render(Primer::AutoComplete.new(src: "/users/search", id: "user-popup", position: :relative)) do |c| %>
   <% c.input(type: :text, name: "input") %>
   <% c.results(classes: "my-custom-class") do %>
     <%= render(Primer::AutoCompleteItemComponent.new(selected: true, value: "value")) do |c| %>
@@ -57,7 +57,7 @@ Use AutoComplete to populate input values from server search results.
 <Example src="<auto-complete src='/users/search' for='user-popup' class='position-relative'>  <input name='input' type='text' class='form-control '></input>  <svg class='octicon octicon-search' height='16' viewBox='0 0 16 16' version='1.1' width='16' aria-hidden='true'><path fill-rule='evenodd' d='M11.5 7a4.499 4.499 0 11-8.998 0A4.499 4.499 0 0111.5 7zm-.82 4.74a6 6 0 111.06-1.06l3.04 3.04a.75.75 0 11-1.06 1.06l-3.04-3.04z'></path></svg>  <ul id='user-popup' class='autocomplete-results '>    <li role='option' data-autocomplete-value='value' aria-selected='true' class='autocomplete-item '>      Selected</li>    <li role='option' data-autocomplete-value='value' class='autocomplete-item '>      Not selected</li></ul></auto-complete>" />
 
 ```erb
-<%= render(Primer::AutoCompleteComponent.new(src: "/users/search", id: "user-popup", position: :relative)) do |c| %>
+<%= render(Primer::AutoComplete.new(src: "/users/search", id: "user-popup", position: :relative)) do |c| %>
   <% c.input(type: :text, name: "input") %>
   <% c.icon(icon: :search) %>
   <% c.results do %>
