+++ test/components/auto_complete_component_test.rb
@@ -2,19 +2,19 @@
 
 require "test_helper"
 
-class PrimerAutoCompleteComponentTest < Minitest::Test
+class PrimerAutoCompleteTest < Minitest::Test
   include Primer::ComponentTestHelpers
 
   def test_raises_if_no_input_is_defined
     assert_raises ArgumentError do
-      render_inline Primer::AutoCompleteComponent.new(src: "/url", id: "my-id") do |component|
+      render_inline Primer::AutoComplete.new(src: "/url", id: "my-id") do |component|
         component.icon(icon: :person)
       end
     end
   end
 
   def test_renders_results_without_explicitly_calling_it
-    render_inline Primer::AutoCompleteComponent.new(src: "/url", id: "my-id") do |component|
+    render_inline Primer::AutoComplete.new(src: "/url", id: "my-id") do |component|
       component.input(name: "input")
     end
 
@@ -25,7 +25,7 @@ class PrimerAutoCompleteComponentTest < Minitest::Test
   end
 
   def test_renders_with_icon
-    render_inline Primer::AutoCompleteComponent.new(src: "/url", id: "my-id") do |component|
+    render_inline Primer::AutoComplete.new(src: "/url", id: "my-id") do |component|
       component.input(name: "input")
       component.icon(icon: :person)
     end
@@ -38,7 +38,7 @@ class PrimerAutoCompleteComponentTest < Minitest::Test
   end
 
   def test_renders_results_with_custom_classes
-    render_inline Primer::AutoCompleteComponent.new(src: "/url", id: "my-id") do |component|
+    render_inline Primer::AutoComplete.new(src: "/url", id: "my-id") do |component|
       component.input(classes: "custom-class")
       component.results(classes: "my-class")
     end
